Convert Global Search from route component to modal.
