+++ src/modules/search/GlobalSearch.scss
@@ -1,0 +1,11 @@
+.mantine-Accordion-content.filters-content {
+  padding: 0;
+}
+.mantine-Accordion-item.filters-item {
+  border-bottom: none;
+}
+.mantine-Accordion-label.filters-label {
+  padding-top: 0.25rem;
+  padding-bottom: 0.25rem;
+  border-radius: 0.25rem;
+}
